requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.26.0
tiktoken
openai==0.23.1
python-docx
python-pptx
pandas
striprtf
nltk
fuzzywuzzy
PyMuPDF
scikit-learn
urllib3
ipaddress
python-Levenshtein